feat: Too much to write in one sentence so just one word Auth
As mentioned above there is too much to write but this was a pretty good experience and fun.
Wrote an authentication from scratch after a while imprinted an understanding of why we create user with unique email
and how to navigate the organization and workspace architecture such that we can add any user to any organization and workspace.
This PR lays the groundwork for that where we will introduce workspace and membership for enterprise offering.
Ask me why I am thinking about enterprise while building an oss project. Then how will I survive as a startup to create a great charting product we need resource and
we will get resource from our enterprise offering.